Updates and revisions
Latest revisions to Polynomial.java
It can now evaluate a polynomial with any given value.  However, the problem still remains if there is a term with a coefficient of one and it does not recognize this is one.  It only sees it as an empty list which creates a problem.  This goes for consecutive terms as well.

Problems being worked on currently.